Add final ints for data amounts, addLabel method, scanImages Refactor
The code that constructed the relative file path (in scanImages()) for
the corresponding data file being used was separated into it's own
method, getFilePath(). The same was done for the logic identifying the
number of data points in the file: getDataSize().

You can specify how many data points are being used with the final
static int variables at the start of the Machine Learning class.